update crud hotel

+++ backend/src/main/java/com/travel_agent/repositories/HotelRepository.java
@@ -3,6 +3,6 @@ package com.travel_agent.repositories;
 import com.travel_agent.models.entity.hotel.HotelEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HotelRepository extends JpaRepository<HotelEntity, Long>{
-    
+public interface HotelRepository extends JpaRepository<HotelEntity, Integer> {
+
 }
